refactor(circuit): split YAML assembly out of circuit_utils [Clean Code: SRP]
Move to a new `scqubits/core/circuit_yaml_assembly.py`:

- `yaml_like_out_with_pp`
- `assemble_circuit`
- `assemble_transformation_matrix`

`assemble_circuit` and `assemble_transformation_matrix` are advertised
in `scqubits/__init__.py` (top-level public API), so a re-export shim
in `circuit_utils.py` stays. `yaml_like_out_with_pp` is internal-by-
purpose (only used by the other two), so no shim per the user's
"public-by-name but internal-by-purpose" rule.

Drop the shims added in B1c–B1e for symbols that are equally
internal-by-purpose (`matrix_power_sparse`, the three operator
factories, `is_potential_term` / `round_symbolic_expr` /
`keep_terms_for_subsystem`). All internal callers already use the
canonical new modules.

`scqubits/__init__.py`:
- Imports `assemble_circuit` / `assemble_transformation_matrix` from
  the canonical `circuit_yaml_assembly` location.
- Wraps the deferred custom-circuit block in `# isort: off` /
  `# isort: on` markers — isort would otherwise re-alphabetise these
  imports above other modules and trigger a circular import.
